doc, tool: extract parseYAML
