+++ tools/doc/html.js
@@ -14,6 +14,7 @@ const loadGtoc = require('./lib/loadGtoc')
 const toID = require('./lib/toID')
 const render = require('./lib/render')
 const parseLists = require('./lib/parseLists')
+const parseYAML = require('./lib/parseYAML')
 
 module.exports = toHTML;
 
@@ -27,21 +28,6 @@ marked.setOptions({
 });
 
 
-function parseYAML(text) {
-  const meta = common.extractAndParseYAML(text);
-  const html = ['<div class="api_metadata">'];
-
-  if (meta.added) {
-    html.push(`<span>Added in: ${meta.added.join(', ')}</span>`);
-  }
-
-  if (meta.deprecated) {
-    html.push(`<span>Deprecated since: ${meta.deprecated.join(', ')} </span>`);
-  }
-
-  html.push('</div>');
-  return html.join('\n');
-}
 
 
 function linkJsTypeDocs(text) {
